First commit for natty, a new virtual terminal.
Natty  - short for Not A TeleTYpewriter - is a virtual terminal like
xterm, gnome-vte, sh, or rxvt. Unlike these programs, natty is not
intended to emulate a DEC VT-series physical video terminal. Natty is
an experimental project to bring new features to the command-line.

Current command-line tools have stagnated (if you prefer, 'stabilized')
around the ECMA-48/ISO 6429/ANSI X3.64 escape code protocol, which are
defined on the basis of the capabilities of 1980s video terminal
devices. Essentially all terminals in use today are virtual terminals
being run on devices which are significantly more capable than these
machines, and yet the terminal environment has not kept pace with these
developments.

Natty will attempt to remain true to the text-oriented command-line
user interface of the the terminal, while extended it to include new
and more powerful interface features, such as:

    * Full support for rich text formatting.
    * Full and correct support for all of Unicode.
    * Lossless keyboard input.
    * Experimental features such as inline media content, dropdown
      menus, subdividing the character grid, and retaining offscreen
      state.

To achieve these ends, Natty will implement a new and more consistent
escape protocol. This library is an engine for translating that
protocol into screen state changes. It does not directly implement any
means of drawing its state to the graphical display. As of this commit,
natty implements (most) ANSI escape codes, which will be maintained for
backwards compatibility reasons. It does not yet implement the natty
escape codes.

A subdirectory, called anterminal, contains a minimal graphical
terminal, based on GTK/pango/cairo, intended as scaffolding for testing
natty. This terminal is buggy and not feature-rich. It contains two
additional libraries used in this implementation - tty, which wraps the
fork_pty command, and pangocairo, which provides some wrapping and FFI
declarations for pango that are not provided by the gtk-rs project's
pango library.

This implementation of natty is written in Rust.