Move key parsing & modifiers into the input module.
Prior to this commit, key parsing was performed in the datatypes
module. Now it is performed in the input module, because it is a
concern of the input processor and not the Key datatype.

The Modifiers struct is now a private type in the input module instead
of a public type in the datatypes module. This struct is not exchanged
data but an aspect of the input parsing algorithm.