Added the first version of the natty escape codes.
This new escape code protocol is the raison d'etre for this project.
Already, the escape code enables encoding complex actions that cannot
be described with any single ANSI command (though it is not yet capable
of performing all actions that ANSI escape codes describe).

The basic grammar of a natty escape code, as of this commmit, is this:

\E{OPCODE(;ARG;*)({LEN;DATA*)}

    where
        \E      = U+001B, the escape character
        OPCODE  = A hexadecimally encoded integer
        ARG     = One of a handful of semantic data types, formed from
                a sequence of hexademical numbers separated by '.'
        LEN     = A hexademically encoded integer
        DATA    = A sequence of bytes with length equal to LEN.

The OPCODE value determines what specific other components are expected
to follow; the LEN/DATA pairs are referred to as "attachments," and are
intended for codes which carry some sort of formatted data with them,
such as the title setting command.

 (Apologies for the informality of this description.)