Natty escape code for writing inline images.
This commit introduces the first natty-exclusive escape code, which
writes image data under the cursor. The command has this structure:

    * OPCODE:   4
    * ARG1:     int, width of image in char cells
                (non-optional)
    * ARG2:     int, height of image in char cells
                (non-optional)
    * ARG3:     media position, how the image will be displayed
                (Default: Display as-is aligned to left and top)
    * ATTACH1:  MIME type for the image format
    * ATTACH2:  Binary data for the image in the defined format

Currently, this command accepts gif, jpeg, and png formatted images.

The testimg.rb file is a ruby script which calls this command. Note
that the terminal must be set not to be in 'onlcr' mode before calling
this command, or the tty system will insert a carriage return before
every newline in the binary data, because computers are terrible. This
can be performed using the command `stty -onlcr`.

This command will frequently send more data than a single read() call
can collect, for this reason the attachment parsing code had to be
modified to handle data larger than is available in the buffer.

Additionally, properly handling image data means not moving the cursor
into cells which would be underneath that image. This was already a
problem for widechars, but the code to skip these char extension cells
was not implemented. The current solution is long-term unacceptable,
but works for now: if the cursor is moving right or downward, it will
be moved to the first rightward cell not within the image area,
otherwise it will be moved to the image's source position. The program
will hang if the image fills the bottom portion of the screen, for
one thing, and this certain motions should perform a totally different
action anyway (newlines for example, should probably move the cursor to
the first line under the image).

The image data is decoded as a part of processing the command, but the
image is not actually drawn to the screen by the scaffolding terminal.
Instead, the scaffolding terminal draws 'X's where the image should be.
This is fine for testing purposes.