Major refactoring to make Input inside Terminal.
Prior to this commit, the Input struct and Terminal struct (formerly
called Screen) were completely separate components. This makes sense
in the context of current terminals, for which user input cannot impact
the screen except by being echoed back from the tty or the controlling
process. But it would not be tenable going forward, because I intend
to implement more stateful terminal behavior, such as local echo or
drop-down menus.

In the new system, the Terminal struct takes ownership of the write
handle to the tty. Two new commands: KeyRelease and KeyPress, should be
constructed from whatever form of keyboard input is made available in
the client library and applied to the Terminal in the same manner as
the commands generated from the Output iterator (e.g. they should all
be sent into a single channel, as they are in the scaffolding
terminal). The Terminal will maintain a consistent state and determine
if any input should be generated from the application of a command.