Drop-down menu tooltips.
A new command creates tooltips which are drop down menus. The menu
contains several options, represented as UTF8 strings, from which the
user can make a selection, navigating with the arrow keys.

Keyboard input which navigates the drop down menu is not transmitted
to the controlling process, but instead is captured. When the user
makes a selection in the menu, a code is transmitted identifying the
selection the user has made.

As of this commit, tooltip op codes are:

ADD TOOL TIP
    OPCODE:     40
    ARG1:       Coords (non-optional)
    ATTACH1:    Tooltip text

ADD DROP DOWN MENU
    OPCODE:     41
    ARG1:       Coords (non-optional)
    ATTACH1-N:  Drop Down Menu options

REMOVE TOOL TIP / DROP DOWN MENU
    OPCODE:     44
    ARG1:       Coords (non-optional)