Move notty esc code types to separate library.
Notty esc code commands and their arguments need to have some form of
existence on both sides of the tty (or whatever system connects their
streams of bytes). For this reason they have been extracted from this
library into a new library called notty-encoding.

Some commands are not implementing using the notty-encoding types; this
is specifically those commands which use attachments. This is because in
the notty implementation, these commands carry their attachments as
RefCell<Option<T>>, rather than T, in order to avoid undue cloning
(commands are only intended to be applied once). Though these commands
have a type in notty-encoding, that type is not used, because this
aspect of their implementation would not make sense in other contexts.