First version of local echo.
In local echo mode, keyboard input causes the direct changes to the
terminal screen.

This commit adds support for echoing some kinds of key input, and for
buffering input in a line-oriented manner. It is similar to the
canonical line discipline in the tty.

A test of this mode, called echotest, is added to the scaffolding
directory. It depends on hocus-pocus, a library for writing
terminal applications.