Replace output parser.
The output parser now operates on an iterator of chars, because all
valid commands are made up of UTF8 encoded code points. This required
major changes throughout the internals of the Output parser.

Attachments are now base64 encoded, instead of raw. This avoids many
problems with transporting binary data over pre-existing terminal
infrastructure, which was designed to translate ASCII/UTF8 data.

Support for Hangul graphemes dropped because it was implemented
incorrectly. This support will be reintroduced correctly in a later
commit.