Use cfg only inside the terminal module.
In order for configurable values to every be defined without
recompilation, these values will have to be stored somewhere, and the
most likely place for that to be is on the Terminal struct. This change
makes this easier by passing cfg::TAB_STOP as an argument to
move_within, so that the only place cfg values are used directly is
inside the terminal module.