Implement Config::update_from_file() method.
This method allows Config struct values to be overridden by values set in a
toml configuration file specified by the "path" parameter. This commit also adds
a test to validate this behavior with a simple config file fixture in the
resources directory. I assume this file will probably need to be moved elsewhere
but will await PR feedback before doing so.