Fix Ctrl-D termination behavior.
Specifically set up a GTK main timeout function to periodically check an
Arc<AtmoicBool> value that indicates whether a read on the pty in the listening
thread fails--this failure is taken to mean that the process on the other side
of the pseudoterminal has exited.

Also improve error handling in CommandApplicator.
Also decrease timeout for CommandApplicatior::apply() calls to improve
responsiveness to keystrokes (notice difference between 125 and 25 milliseconds
when holding down a character in scaffolding)