Make Command more opaque.
In order to eventually make `Terminal` an opaque type, commands need to
be passed to a terminal method, instead of calling methods on terminal
publicly. Additionally, proper local echo support will require a certain
consistent dispatch when commands are applied, unless they are applied
internally. This refactor changes no behavior, but is a step toward
reducing the public API of the notty library.