Subdividable screens.
Introduce a layer of indirection between `Terminal` and `CharGrid`,
using the `Screen` type, which allows a `Terminal` to contain multiple
`CharGrids` in different sections of the visible screen.

Prior to this commit, a single `Terminal` contained a stack of
`CharGrid`s, which each filled the entire size of the `Terminal` window.

Now, a single `Terminal` contains a single `Screen`, which is divided
into `ScreenSection`s. Each `ScreenSection` is a rectangular section of
the screen. They are formed by splitting an existing section into two
along a horizontal or vertical axis. This means that the relationship
between sections that exist is always a binary tree. Each
`ScreenSection` has an unsigned 64-bit tag. One section is identified as
the 'active section' - commands that apply to the `CharGrid` are applied
to the grid contained in this section.

Each section contains a non-empty stack of `Panel`s. A `Panel` can take
one of two forms: either it contains a `CharGrid`, or it is split into
two parts, each of which contains a `ScreenSection`. As a result, there
is a mutually recursive relationship between the definition of
`ScreenSection` and `Panel`, which allows for considerable flexibility
in adjusting the layout screen, while still maintaining a certain level
of restraint.

This commit and its associated commit to notty-encoding add or modify
several commands to the notty protocol:

* __PUSH BUFFER__ is now __PUSH PANEL__, and takes as an argument the
tag of the screen section being pushed to. If no argument is passed, a
panel is pushed to the active section. The panel that is pushed always
contains a single character grid. It also takes a boolean argument,
which determines whether or not the grid retains offscreen data as it
scrolls.
* __POP BUFFER__ is now __POP PANEL__, and also takes as an argument the
tag of screen section to pop from. If no argument is passed, the top
panel of the active section is popped. A stack with 1 member cannot be
popped from.
* __SPLIT PANEL__ describes splitting a panel into two. It takes
arguments defining the axis and position of the split, which panel the
contents should be saved to, how resizing of child sections should be
handled, and whether or not the grid in the new panel should retain
offscreen data as it scrolls. Panels which are already split *can* be
split again; the existing split sections are resized into the
identified saved sections.
* __UNSPLIT PANEL__ describes removing the split from a panel. It takes
an argument identifying which of the two sections in this panel should
be saved. The stack of the saved section will be pushed to the top of
the section this panel is on top of; the stack of the unsaved section
will be destroyed.
* __ADJUST PANEL SPLIT__ allows an existing split panel's split to be
moved or adjusted. It takes arguments defining the axis and position of
the split, and how resizing of the child sections should be handled. It
performs no action it the section's top is a unsplit panel.
* __ROTATE SECTION DOWN__ rotates down the stack, putting the top of
the stack on the bottom.
* __ROTATE SECTION UP__ rotates up the stack, putting the bottom of the
stack on the top.
* __SWITCH ACTIVE SECTION__ switches which section is active. It will
only perform a switch if the top panel of the target section is a single
character grid.