Add chars_in_range to CharGrid.
This function returns a String of all the characters in the cells from
one coordinate point to another, include, moving right and then down. It
does not include any non-character data, and empty cells are not
represented in the output. If the coordinates wrap around a linebreak, a
newline character is included there.

The purpose of this function is for selecting text on the screen to move
into the system clipboard.