Apply color configuration in the rendering layer.
Rather than translating a default or palette color to a true color when
it is applied at a style, preserve that information and translate it to
a true color when it is rendered.

This makes the notty library not dependent on color configuration, and
paves the way for changing default and palette configurations in an app
without restarting.