Preconfigured styles for syntax group.
This feature allows the user to set preconfigured styles associated with
different syntactic components of the text on the screen. Currently, it
supports the 'plain' text configuration, which will be used by default,
as well as configuration for 10 different 'code groups', relevant when
editing source code. It will be expanded to more syntax groups in the
future.

In the long term, the goal is to get programs to migrate to using these
syntax groups instead of manually setting their own styles, so that
users can have a consistent and easily configurable experience across
different terminal applications.

In scaffolding, syntax groups can be styled in the config file.